configuration: simplifying type usage

+++ himan-bin/source/himan.cpp
@@ -27,13 +27,13 @@ shared_ptr<configuration> ParseCommandLine(int argc, char** argv);
 struct plugin_timing
 {
 	std::string plugin_name;
-	unsigned short order_number;  // plugin order number (if called more than once))
+	int order_number;  // plugin order number (if called more than once))
 	size_t time_elapsed;          // elapsed time in ms
 };
 
-unsigned short HighestOrderNumber(const vector<plugin_timing>& timingList, const std::string& pluginName)
+int HighestOrderNumber(const vector<plugin_timing>& timingList, const std::string& pluginName)
 {
-	unsigned short highest = 1;
+	int highest = 1;
 
 	for (size_t i = 0; i < timingList.size(); i++)
 	{
@@ -41,7 +41,7 @@ unsigned short HighestOrderNumber(const vector<plugin_timing>& timingList, const
 		{
 			if (timingList[i].order_number >= highest)
 			{
-				highest = static_cast<unsigned short>(timingList[i].order_number + 1);
+				highest = timingList[i].order_number + 1;
 			}
 		}
 	}
@@ -475,7 +475,7 @@ shared_ptr<configuration> ParseCommandLine(int argc, char** argv)
 		conf->UseCudaForUnpacking(false);
 	}
 
-	conf->CudaDeviceCount(static_cast<short>(devCount));
+	conf->CudaDeviceCount(devCount);
 
 	if (opt.count("cuda-device-id"))
 	{
